Resolving conflicts with testing
Jesse writes:
ethanrowe:tickets/master/2239 leaks state in the spec. After the spec
is run, Puppet::Application is left in a :restart_requested state, and
several important behaviors, particularly Puppet::Transaction#evaluate
are disabled.

It's order dependent, so changing the mtime of spec files makes the
failures appear and disappear.

This spec file was generally pretty good about keeping state from
getting out, but one test was missing a stub for a dangerous call.

I wouldn't be surprised if this cleared up other errors in testing.