Storing per-environment LoadedCode instances
This will soon replace all of the env/parser mungling
we have to do.  A given process will only be able to
have one collection of code per environment in memory.

This is somewhat limiting, in theory, but some global means
of looking up code collection (LoadedCode instances) must
exist for the pure ruby stuff to work.

Signed-off-by: Luke Kanies <luke@reductivelabs.com>