[#4256] External nodes parameters can now be assigned to nodes
Node parameters were made a reader instead of an accessor in

  commit b82b4ef04282ca0006931562f60459a1591b6268
  Author: Luke Kanies <luke@reductivelabs.com>
  Date:   Wed Jan 6 17:42:42 2010 -0800

      All non-transient parser references are gone

but external nodes needs to be able to assign to parameters.  The fix is
just to change that back to an accessor.  There may have been concern
over nodes replacing the hash object instead of the values could have
bad consequences, but that's not a concern since the node object being
created in this case is new also.

Paired with:  Nick Lewis