Fix for #4506 -- too much data being serialized
The serialization of much of the memory image was being triggered by a copy of
an unchanging method return value in an instance variable.  I had introduced
this as a performance hack.  The results were correct, but when the object was
serialized it took a huge amount of data with it (known_resource_types, and
from there...)

As the underlying costly operation is no longer being called (thanks to Paul and
Jesse) the caching is no longer a significant performance boost, and can thus
simply be removed.